Added temperature to the Weather class, added temperature to the customer class, try to run Game

+++ LemonadeStand/Ingredient.cs
@@ -9,7 +9,7 @@ namespace LemonadeStand
     {
         //member variable
         public string name;
-        private int price;
+        private decimal price;
 
 
         //constructor
